+++ tests/fixtures/middleware-included-paths/middleware-node.ts
@@ -1,0 +1,6 @@
+export { middleware } from './middleware-shared'
+
+export const config = {
+  matcher: ['/test', '/link'],
+  runtime: 'nodejs',
+}